+++ projects/pip-webui2-dates/src/lib/shared/pipes/index.ts
@@ -1,0 +1,1 @@
+export { PipDatePipesModule } from './pipes.module';