+++ projects/pip-webui2-dates/src/lib/shared/pipes/pipes.module.ts
@@ -1,0 +1,17 @@
+import { NgModule } from '@angular/core';
+
+import { LocalizedDatePipe } from './localized-date.pipe';
+import { ElapsedTimePipe } from './elapsed-date.pipe';
+
+const PIPES = [
+    LocalizedDatePipe,
+    ElapsedTimePipe
+];
+
+@NgModule({
+    imports: [],
+    declarations: [...PIPES],
+    providers: [],
+    exports: [...PIPES]
+})
+export class PipDatePipesModule { }