+++ projects/pip-webui2-dates/src/lib/shared/translations/elapsed-pipe.translations.ts
@@ -1,0 +1,18 @@
+export let ElapsedPipeTranslations = {
+    'en': {
+        'DATES.MOMENT_AGO': 'just moment ago',
+        'DATES.SECONDS_AGO': '{{ value }} sec ago',
+        'DATES.MINUTES_AGO': '{{ value }} min ago',
+        'DATES.HOURS_AGO': '{{ value }} hour(s) ago',
+        'DATES.DAYS_AGO': '{{ value }} day(s) ago',
+        'DATES.WEEKS_AGO': '{{ value }} week(s) ago'
+    },
+    'ru': {
+        'DATES.MOMENT_AGO': 'только что',
+        'DATES.SECONDS_AGO': '{{ value }} сек назад',
+        'DATES.MINUTES_AGO': '{{ value }} мин назад',
+        'DATES.HOURS_AGO': '{{ value }} час назад',
+        'DATES.DAYS_AGO': '{{ value }} дн назад',
+        'DATES.WEEKS_AGO': '{{ value }} нед назад'
+    }
+};