+++ projects/pip-webui2-dates/src/lib/time-range-edit/index.ts
@@ -1,0 +1,2 @@
+export { PipTimeRangeEditComponent } from './time-range-edit.component';
+export { PipTimeRangeEditModule } from './time-range-edit.module';