+++ projects/pip-webui2-dates/src/lib/time-range-edit/shared/time-range-edit.translations.ts
@@ -1,0 +1,10 @@
+export const TimeRangeEditTranslations = {
+    en: {
+        'DATES.START_TIME': 'Start time',
+        'DATES.END_TIME': 'End time'
+    },
+    ru: {
+        'DATES.START_TIME': 'Время начала',
+        'DATES.END_TIME': 'Время окончания'
+    }
+};