+++ projects/pip-webui2-dates/src/lib/time-range/index.ts
@@ -1,0 +1,2 @@
+export { PipTimeRangeComponent } from './time-range.component';
+export { PipTimeRangeModule } from './time-range.module';