+++ src/app/time-range-edit-example/time-range-edit-example.module.spec.ts
@@ -1,0 +1,13 @@
+import { TimeRangeEditExampleModule } from './time-range-edit-example.module';
+
+describe('TimeRangeEditExampleModule', () => {
+  let timeRangeEditExampleModule: TimeRangeEditExampleModule;
+
+  beforeEach(() => {
+    timeRangeEditExampleModule = new TimeRangeEditExampleModule();
+  });
+
+  it('should create an instance', () => {
+    expect(timeRangeEditExampleModule).toBeTruthy();
+  });
+});