+++ src/app/time-range-example/time-range-example.module.spec.ts
@@ -1,0 +1,13 @@
+import { TimeRangeExampleModule } from './time-range-example.module';
+
+describe('TimeRangeExampleModule', () => {
+  let timeRangeExampleModule: TimeRangeExampleModule;
+
+  beforeEach(() => {
+    timeRangeExampleModule = new TimeRangeExampleModule();
+  });
+
+  it('should create an instance', () => {
+    expect(timeRangeExampleModule).toBeTruthy();
+  });
+});