+++ src/index.ts
@@ -1,3 +1,0 @@
-export * from './time-range/index';
-export * from './time-range-edit/index';
-export * from './shared/pipes/index';