+++ src/shared/pipes/index.ts
@@ -1,1 +1,0 @@
-export { PipDatePipesModule } from './pipes.module';