+++ src/time-range-edit/index.ts
@@ -1,2 +1,0 @@
-export { PipTimeRangeEditComponent } from './time-range-edit.component';
-export { PipTimeRangeEditModule } from './time-range-edit.module';