+++ src/time-range/index.ts
@@ -1,2 +1,0 @@
-export { PipTimeRangeComponent } from './time-range.component';
-export { PipTimeRangeModule } from './time-range.module';