+++ CHANGELOG.md
@@ -14,3 +14,18 @@ No breaking changes in this version
 
 ### Bug Fixes
 No fixes in this version 
+
+## <a name="0.0.2"></a> 0.0.2 (2018-10-02)
+
+Update to Angular 6
+
+### Features
+* **date time**: date and time editing controls
+* **localized date**: pipe that shows localized date
+* **elapsed date**: pipe that shows localized date range from date to now
+
+### Breaking Changes
+No breaking changes in this version
+
+### Bug Fixes
+No fixes in this version